Add a 'has' method to the EnumDescriptor class to make it easy to check whether a named constant exists in the enumeration without having to catch the error on the get methods.
Also, fix the length method for EnumDescriptor which was broken due to
a typo in the C++ method name that the R tried to invoke.
         
Update the show method for EnumDescriptors to display how many named
constants are in the enumeration.

Add a number method to EnumValueDescriptors to return the numeric
value of this enum constant.



git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rprotobuf@494 edb9625f-4e0d-4859-8d74-9fd3b1da38cb