Add a large new section showing how RProtoBuf can be used to serialize all of the example datasets that ship with R, and compare the sizes with R's built-in serialziation method.
Steal a section on applications/MapReduce from the HistogramTools
vignette.  To be improved further.

Combine each of the class slot and method tables into a single table
for each class.

Add a brief section on type coercion issues for e.g. bools and int64s.

Improve the wording in various places.



git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rprotobuf@624 edb9625f-4e0d-4859-8d74-9fd3b1da38cb