Section 5 (types): Add a sentence noting that three types in particular need special attention, to transition to the three subsections on booleans, uint32s, and int64s.
Section 6 (Converting R data structures to Protocol Buffers):
Rewrite first two sentences to make the distinction between the
previous sections describing methods for working with messages and a
specific pre-defined schema, and the function in this section
describing a universal schema that can be used on arbitrary R data
structures.  This seems like the more accurate distinction than
calling it low-level vs high-level as was done previously.

Also, add a missing comma in section 6.



git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rprotobuf@774 edb9625f-4e0d-4859-8d74-9fd3b1da38cb