Add a short sentence to define serialization early in the intro, addressing a todo that multiple people had mentioned.  Reference the C++ FAQ for lack of a better reference for now.
Revert the second to last paragraph of the introduction to my earlier
version that was instead moved to section 2.  Remove one of the more
technical sentences to address Jeroen's observation that it was a bit
too technical for the intro (e.g. "reflection" and dynamic typed
languages was a bit much)

Most sentences of the deleted paragraph were false, as discussed in
email.



git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rprotobuf@802 edb9625f-4e0d-4859-8d74-9fd3b1da38cb