Mostly work on section 2 to address some flaws identified by Jeroen, Karl, and others.
Move up the basic description of the protocol buffer schema from
section 3 to section 2, including the example of how protocol buffers
are manipulated with this package.

Revert a regression -- fix the reference to BSON and MessagePack by
putting the citations next to the text about the R interfaces rather
than the formats themselves (re-apply fix from Dirk).

Add an explicit transition from section 2 to section 3 as the last sentence of 2.

Define payload at the beginning of section 3 just once, so we don't
repeat ourselves later in the section.

Add a sentence to section 6 that provides more context about when you
use the basic RProtoBuf functionality with specific schemas -- "This
is useful when there are pre-existing systems with defined schemas or
significant software components written in other languages that need
to be accessed from within R." before transitioning to talk about the
universal r language schema.  (This point suggested by Karl).



git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rprotobuf@805 edb9625f-4e0d-4859-8d74-9fd3b1da38cb