Add back a reference to Tierney's serialization doc.
Move my incomplete TODO performance summary subsection to the final
conclusion/summary section, where I will likely delete most of it but
may use a bit of those ideas in wrapping everything up in a
conclusion.  Ditto for the commented out other approaches section.

Section 6 previously ended on a pessimistic note and a sentence
fragment (the table shows protobuf doesn't do much better in
compression size than normal R serialization).

End on a more complete note that mentions that RProtoBuf is most
benefetial when multiple languages are involved, and when a more
concise application-specific schema is in place, and transition to the
example in the next section by noting that both of those conditions
hold for the MapReduce / histogram example.



git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rprotobuf@808 edb9625f-4e0d-4859-8d74-9fd3b1da38cb