Updates to section 7:
Add a better transition to start section 7 reminding the user what the
application in section 6 was about and how/why this one is different.

Remove several duplicate sentences about the basics of protocol buffer
.proto files and such which are explained earlier in the paper.

Remove a few sentences that provide unnecessary level of detail about
OpenCPU.  In this section it is an example web service and so we don't
need to advertise to the reader its other capabilities that are
unnecessary for the example.

Backreference to the section about serialize_pb here.

Avoid saying 'protobuf' or 'protobuf messages' as this terminology was
only used in this section.  Instead, spell out protocol buffers.

End result is 13-lines shorter/more concise (1/3 - 1/2 of a page), and
I think clearer as well.

We still don't mention the word RESTful anywhere here, where OpenCPU
is a RESTful web service where just one argument of the POST request
is encoded with the protocol buffer, instead of a more general non
web/REST type of RPC server that would tend to be a more natural fit
for protocol buffers.



git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rprotobuf@910 edb9625f-4e0d-4859-8d74-9fd3b1da38cb