Cut section 6 in half by removing the section explaining the caveats about formulas and types not supported by serialized_pb, and just explain now that we serialize everything, but in one sentence explain the caveat that we fall back to base::serialize for R-specific types like language,function,and environment.  This basically removes the need for 6.1 at all, so remove that section, and move a tiny bit of the top text about which datasets we are using into the top of section 6.2 which explains the compresison performance.
Next step: Replace table with a plot as hadley and one of the referees
both suggested.

Another referee suggested just merging 5 and 6 together completely.
Now that section 6 is one page plus one large table, that seems
feasible, but I'll revisit that after replacing the table with a plot.
I do want to make a stark distinction between having a schema (section
5) and not (section 6).  I never really use section 6 schema-less
method, but its the one that is easier for people to play around with
probably if they don't have a real application with protocol buffers
yet.



git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rprotobuf@914 edb9625f-4e0d-4859-8d74-9fd3b1da38cb